Week-2/Day-2/full_parameter_finetuning/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bardapi==0.1.38
bcrypt==4.1.3
beautifulsoup4==4.12.2
bleach==6.2.0
blinker==1.8.2
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.3.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.14
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.19.1
Flask==3.0.3
fqdn==1.5.1
fsspec==2025.7.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-translate==3.14.0
google-genai==1.31.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.1.0
hf-xet==1.1.9
hpack==4.0.0
httpcore==1.0.2
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
langdetect==1.0.9
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
numpy==1.26.3
ollama==0.5.3
openai==1.72.0
overrides==7.7.0
packaging==25.0
pandas==2.1.4
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.51
proto-plus==1.23.0
protobuf==4.25.2
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.11.3
pydantic_core==2.33.1
pyftpdlib==1.5.10
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2023.3.post1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
safetensors==0.6.2
Send2Trash==1.8.3
setuptools==79.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.14
